Add missing licenses
