+++ src/matter/cluster/server/ClusterServer.ts
@@ -1,3 +1,9 @@
+/**
+ * @license
+ * Copyright 2022 Marco Fucci di Napoli (mfucci@gmail.com)
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 import { MatterDevice } from "../../MatterDevice";
 import { Session } from "../../session/Session";
 import { AttributeSpecs, ClusterSpec, CommandSpec, CommandSpecs, TypeFromAttributeSpec } from "../ClusterSpec";
