+++ src/matter/cluster/server/GeneralCommissioningServer.ts
@@ -1,3 +1,9 @@
+/**
+ * @license
+ * Copyright 2022 Marco Fucci di Napoli (mfucci@gmail.com)
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 import { MatterDevice } from "../../MatterDevice";
 import { SecureSession } from "../../session/SecureSession";
 import { CommissioningError, GeneralCommissioningClusterSpec } from "../GeneralCommissioningCluster";
